// backend/src/modules/stock/enums/stock-movement-type.enum.ts
export enum StockMovementType {
  INITIAL = 'INITIAL',
  INCREMENT = 'INCREMENT',
  DECREMENT = 'DECREMENT',
  MINIMUM_CHANGE = 'MINIMUM_CHANGE',
  DEACTIVATION = 'DEACTIVATION',
  REACTIVATION = 'REACTIVATION',
  PURCHASE = 'PURCHASE',
  OPENING_VALUATION = 'OPENING_VALUATION',
  PURCHASE_CANCELLATION = 'PURCHASE_CANCELLATION',
  VALUATION_CANCELLATION = 'VALUATION_CANCELLATION',
}